Separate unit and integration tests into different folders
