+++ test/unit/middleware/api-documentation.test.ts
@@ -1,8 +1,8 @@
 import { Response } from 'koa';
 import parseLinkHeader from 'parse-link-header';
-import apiDocumentation from '../../src/middleware/api-documentation-link';
-import createContext from '../context';
-import runMiddleware, { NextMiddleware } from '../middleware';
+import apiDocumentation from '../../../src/middleware/api-documentation-link';
+import createContext from '../../context';
+import runMiddleware, { NextMiddleware } from '../../middleware';
 
 const makeRequest = async (next?: NextMiddleware): Promise<Response> => (
   runMiddleware(apiDocumentation('/path-to/api-documentation'), createContext(), next)
