+++ test/unit/middleware/error-handler.test.ts
@@ -3,11 +3,11 @@ import createHttpError from 'http-errors';
 import { INTERNAL_SERVER_ERROR, SERVICE_UNAVAILABLE } from 'http-status-codes';
 import 'jest-rdf';
 import { Response } from 'koa';
-import { WithDataset } from '../../src/middleware/dataset';
-import errorHandler from '../../src/middleware/error-handler';
-import { hydra, rdf } from '../../src/namespaces';
-import createContext, { ErrorListener } from '../context';
-import runMiddleware, { NextMiddleware, throwingNext } from '../middleware';
+import { WithDataset } from '../../../src/middleware/dataset';
+import errorHandler from '../../../src/middleware/error-handler';
+import { hydra, rdf } from '../../../src/namespaces';
+import createContext, { ErrorListener } from '../../context';
+import runMiddleware, { NextMiddleware, throwingNext } from '../../middleware';
 
 const makeRequest = async (next?: NextMiddleware, errorListener?: ErrorListener): Promise<WithDataset<Response>> => (
   runMiddleware(errorHandler(), createContext({ errorListener }), next)
