+++ test/unit/middleware/jsonld.test.ts
@@ -6,11 +6,11 @@ import 'jest-rdf';
 import { Context as JsonLdContext } from 'jsonld/jsonld-spec';
 import { addAll } from 'rdf-dataset-ext';
 import { Quad } from 'rdf-js';
-import { AppContext } from '../../src/app';
-import jsonld from '../../src/middleware/jsonld';
-import { rdf, schema } from '../../src/namespaces';
-import createContext, { Headers } from '../context';
-import { NextMiddleware } from '../middleware';
+import { AppContext } from '../../../src/app';
+import jsonld from '../../../src/middleware/jsonld';
+import { rdf, schema } from '../../../src/namespaces';
+import createContext, { Headers } from '../../context';
+import { NextMiddleware } from '../../middleware';
 
 const makeRequest = async (
   body?: string, headers?: Headers, next: NextMiddleware = jest.fn(), jsonLdContext: JsonLdContext = {},
