+++ test/unit/router.test.ts
@@ -1,5 +1,5 @@
-import createRouter from '../src/router';
-import Routes from '../src/routes';
+import createRouter from '../../src/router';
+import Routes from '../../src/routes';
 
 describe('router', (): void => {
   const router = createRouter();
