+++ test/unit/routes/add-article.test.ts
@@ -5,13 +5,13 @@ import all from 'it-all';
 import 'jest-rdf';
 import { Response } from 'koa';
 import { DatasetCore } from 'rdf-js';
-import InMemoryArticles from '../../src/adaptors/in-memory-articles';
-import Articles from '../../src/articles';
-import { schema } from '../../src/namespaces';
-import addArticle from '../../src/routes/add-article';
-import createContext from '../context';
-import createArticle from '../create-article';
-import runMiddleware, { NextMiddleware } from '../middleware';
+import InMemoryArticles from '../../../src/adaptors/in-memory-articles';
+import Articles from '../../../src/articles';
+import { schema } from '../../../src/namespaces';
+import addArticle from '../../../src/routes/add-article';
+import createContext from '../../context';
+import createArticle from '../../create-article';
+import runMiddleware, { NextMiddleware } from '../../middleware';
 
 const makeRequest = async (
   dataset?: DatasetCore,
